Set DOTNET_ROLL_FORWARD for tests
Set `DOTNET_ROLL_FORWARD` for `dotnet test` when upgrading to a preview to fix issue that can cause .NET tools to not be able to use the preview version of the runtime if a better version isn't available.
For example, Microsoft.Extensions.ApiDescription.Server for .NET 9 previews only targets net9.0 and netcoreapp2.1, and without roll-forward it trues to use .NET Core 2.1, which isn't really installed anywhere anymore as it is out-of-support.